test: decompose plot-plan-runs.test.ts below 500-line limit (warren-59db)
Extract the file-local helper/fixture group into a new sibling
src/server/handlers/plot-plan-runs.test-helpers.ts (one copy of every
stub/fixture, jscpd-scanned). Split the single describe block's 9 tests
into plot-plan-runs.test.ts (happy-path + filter) and a new
plot-plan-runs.validation.test.ts (7 validation/error tests), each
importing from the helpers module. Test bodies, assertions, stubs, and
fixtures are unchanged; 9 pass / 0 fail.

Mirrors checks.test-helpers.ts (warren-7a15) and
projects.test-helpers.ts (warren-a715).
